Move sync-dropdown component to staging-site-sync-dropdown directory
